requirements.txt
fastapi
uvicorn
python-dotenv
httpx
sentence-transformers
qdrant-client
pydantic
